Add a simple visualizer for sanity checks.
This uses Python and Graphviz.  It's quick to catch typos since they'll
appear as leaf nodes.  But graphviz graps aren't meant to be used as
data structures, so it's not sufficient.

I don't want to invest too much effort in making the graphviz graph
look good when I'll still need to make something else to produce the
final materials list.